backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
boto3==1.38.10
botocore==1.38.10
certifi==2025.4.26
charset-normalizer==3.4.2
exceptiongroup==1.2.2
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-aws==0.2.22
langchain-core==0.3.58
langchain-text-splitters==0.3.8
langsmith==0.3.42
numpy==1.26.4
orjson==3.10.18
packaging==24.2
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
zstandard==0.23.0
snowflake-connector-python==3.15.0
cryptography==44.0.1
pandas==2.2.2
fastapi
uvicorn
matplotlib==3.8.2
seaborn==0.13.0
pandasql
python-multipart
